% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hef.R
\name{hef}
\alias{hef}
\title{Hierarchical Exponential Family Model}
\usage{
hef(model = c("bin_beta"), data, prior = NULL, hpars = NULL, ...)
}
\arguments{
\item{model}{A character string.  Abbreviated name for the
response-mixture combination.}

\item{data}{A numeric matrix.  The format depends on \code{model}.}

\item{prior}{Optional.  If the user wishes to specify their own prior
then \code{prior} should be an R function that has the arguments
\code{x} and \code{hpars} and returns the log of the prior.
The argument \code{x} is the ...
list of prior function and hyperparemeters
Provide a convenience function.}

\item{hpars}{A vector of hyperparameter values.  Default only?}

\item{...}{Further argments to be passed to \code{\link[rust]{ru}}.  Most
importantly \code{trans} and \code{rotate} (see \strong{Details}), and
perhaps \code{r}, \code{ep}, \code{a_algor}, \code{b_algor},
\code{a_method}, \code{b_method}, \code{a_control}, \code{b_control}.
May also be used to pass the arguments \code{n_grid} and/or \code{ep_bc}
to \code{\link[rust]{find_lambda}}.}
}
\description{
Samples from the posterior distribution of hierarchical models in
which the response distribution is in the exponential family
with a canonical mixing distribution.
The user can either choose hyperparameter values of a default prior
distribution or specify their own prior distribution.
}
\examples{
bb_data <- sim_bin_beta(n = 10, size = 20, alpha = 2, beta = 2)
bb_res <- hef(model = "bin_beta", data = bb_data)
plot(bb_res)

}
