% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{iid}
\alias{iid}
\title{Bayesian Inference for Univariate Distributions}
\usage{
iid(x, densfun, start, nsim = 1000, prior = "default", hpars = NULL,
  param = "trans", lambda = NULL, gamma_rate = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector of length at least one containing only
\code{\link{finite}} values}

\item{densfun}{Either a character string or a function returning a density
  evaluated at its first argument.

  Distributions \code{"beta"}, \code{"cauchy"}, \code{"chi-squared"},
  \code{"exponential"}, \code{"f"}, \code{"gamma"}, \code{"geometric"},
  \code{"log-normal"}, \code{"lognormal"}, \code{"logistic"},
  \code{"negative binomial"}, \code{"normal"}, \code{"Poisson"},
  \code{"t"} and \code{"weibull"} are recognised, case being ignored.}

\item{start}{A named list giving the parameters to be optimized with
initial values. This can be omitted for some of the named distributions
and must be for others (see \strong{Details}).}

\item{nsim}{A numeric scalar. The number of values to be simulated from the
posterior distribution.}

\item{prior}{Describe.}

\item{hpars}{Describe.}

\item{param}{Describe.}

\item{lambda}{A numeric vector.  rep_len}

\item{gamma_rate}{A logical scalar.}

\item{...}{Additional parameters, either for \code{densfun} or
\code{\link[rust]{ru}}.
In particular, it can be used to specify bounds via lower
or upper or both. If arguments of densfun (or the density function
corresponding to a character-string specification) are included they will
be held fixed.}
}
\value{
Add return
}
\description{
Similar to fitdistr
}
\details{
Add details
}
\examples{
x <- rgeom(10, 0.5)
# fitdistr(x, "geometric")
geom_prior <- set_user_prior(dbeta, shape1 = 1, shape2 = 1, log = TRUE,
                             model = "iid", par_names = "prob")
set.seed(47)
pjn1 <- iid(x, "geometric", prior = geom_prior)
set.seed(47)
pjn2 <- iid(x, "geometric", prior = "beta")
set.seed(47)
pjn3 <- iid(x, "geometric", prior = "default")
}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied
  Statistics with S.  Fourth Edition. Springer, New York.
  ISBN 0-387-95457-0. \url{http://www.stats.ox.ac.uk/pub/MASS4}.
}
\seealso{
The \code{\link[rust]{ru}} function in the \code{\link{rust}}
  package for details of the arguments that can be passed to \code{ru}.

\code{\link{set_user_prior}} to set a user-defined prior.
}
