% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesplot.R
\name{pp_check.hef}
\alias{pp_check.hef}
\alias{pp_check}
\title{Posterior predictive checks for a hef object}
\usage{
\method{pp_check}{hef}(object, fun = NULL, raw = FALSE, nrep = NULL, ...)
}
\arguments{
\item{object}{An object of class "hef", a result of a call to
\code{\link{hef}} or \code{\link{hanova1}}.}

\item{fun}{The plotting function to call.
Can be any of the functions detailed at \link[bayesplot]{PPC-overview}.
The "ppc_" prefix can optionally be dropped if fun is specified
as a string.}

\item{raw}{Only relevant if \code{object$model = "beta_binom"} or
\code{object$model = "gamma_pois"}.
If \code{raw = TRUE} then the raw responses are used in the plots.
Otherwise, the \emph{proportions} of successes are used in the
\code{beta_binom} case and the \emph{exposure-adjusted rate} in the
\code{gamma_pois} case.  In both cases the values used are
 \code{object$data[, 1] / object$data[, 2]} and the equivalent
 in \code{object$data_rep}.}

\item{nrep}{The number of predictive replicates to use.  If \code{nrep}
is supplied then the first \code{nrep} rows of \code{object$data_rep}
are used.  Otherwise, or if \code{nrep} is greater than
\code{nrow(object$data_rep)}, then all rows are used.}

\item{...}{Additional arguments passed on to bayesplot functions.
See \strong{Examples} below.}
}
\value{
A ggplot object that can be further customized using the
  \strong{ggplot2} package.
}
\description{
\code{pp_check} method for class "hef".  This provides an interface
to the functions that perform posterior predictive checks in the
\strong{bayesplot} package.  See \link[bayesplot]{PPC-overview} for
details of these functions.
}
\details{
For details of these functions see \link[bayesplot]{PPC-overview}.
  See also the vignettes
  \href{https://CRAN.R-project.org/package=bang}{Conjugate Hierarchical Models},
  \href{https://CRAN.R-project.org/package=bang}{Hierarchical 1-way Analysis of Variance}
  and the \strong{bayesplot} vignette
  \href{https://CRAN.R-project.org/package=bayesplot}{Graphical posterior predictive checks}.

  The general idea is to compare the observed data \code{object$data}
  with a matrix \code{object$data_rep} in which each row is a
  replication of the observed data simulated from the posterior predictive
  distribution.  For greater detail see Chapter 6 of
  \href{www.stat.columbia.edu/~gelman/book}{Gelman et al. (2014)}.
}
\examples{
############################ Beta-binomial #################################

# ------------------------- Rat tumor data ------------------------------- #

rat_res <- hef(model = "beta_binom", data = rat, nrep = 50)

# Overlaid density estimates
pp_check(rat_res)
# Overlaid density estimates
pp_check(rat_res, fun = "ecdf_overlay")
# Multiple histograms
pp_check(rat_res, fun = "hist", nrep = 8)
# Multiple boxplots
pp_check(rat_res, fun = "boxplot")
# Predictive medians vs observed median
pp_check(rat_res, fun = "stat", stat = "median")
# Predictive (mean, sd) vs observed (mean, sd)
pp_check(rat_res, fun = "stat_2d", stat = c("mean", "sd"))

############################ Gamma-Poisson #################################

# ------------------------ Pump failure data ------------------------------ #

pump_res <- hef(model = "gamma_pois", data = pump, nrep = 50)

# Overlaid density estimates
pp_check(pump_res)
# Predictive (mean, sd) vs observed (mean, sd)
pp_check(rat_res, fun = "stat_2d", stat = c("mean", "sd"))
}
\references{
Jonah Gabry (2016). bayesplot: Plotting for Bayesian
Models. R package version 1.1.0.
\url{https://CRAN.R-project.org/package=bayesplot}

Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B.,
Vehtari, A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis}.
Chapman & Hall/CRC Press, London, third edition. (Chapter 6)
\url{www.stat.columbia.edu/~gelman/book}
}
\seealso{
\code{\link{hef}} and \code{\link{hanova1}} for sampling
  from posterior distributions of hierarchical models.

\strong{bayesplot} functions \link[bayesplot]{PPC-overview},
  \link[bayesplot]{PPC-distributions},
  \link[bayesplot]{PPC-test-statistics},
  \link[bayesplot]{PPC-intervals},
  \link[bayesplot]{pp_check}.
}
