% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diagnostics.R
\name{print.hef}
\alias{print.hef}
\title{Print method for objects of class "hef"}
\usage{
\method{print}{hef}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "hef", a result of a call to
\code{\link{hef}} or \code{\link{hanova1}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}
  methods.
}
\description{
\code{print} method for class "hef".
}
\details{
Prints the original call to \code{\link{hef}} or
  \code{link{hanova1}}, the name of the model and the number of populations
  in the hierarchical model.
}
\seealso{
\code{\link{hef}} for hierarchical exponential family models.

\code{\link{hanova1}} for hierarchical one-way analysis of
  variance (ANOVA).
}
