% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary_print.R
\name{print.summary.hef}
\alias{print.summary.hef}
\title{Print method for objects of class "summary.hef"}
\usage{
\method{print}{summary.hef}(x, ...)
}
\arguments{
\item{x}{an object of class "summary.hef", a result of a call to
\code{\link{summary.hef}}.}

\item{...}{Additional optional arguments to be passed to
\code{\link{print}}.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}
  methods.
}
\description{
\code{print} method for class "summary.hef".
}
\details{
What is printed depends on the argument \code{params} supplied
  to \code{\link{summary.hef}}.
}
\seealso{
\code{\link{summary.hef}}: \code{summary} method for class "hef".

\code{\link{hef}} for hierarchical exponential family models.

\code{\link{hanova1}} for hierarchical one-way analysis of
  variance (ANOVA).
}
