% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bang.R
\docType{data}
\name{pump}
\alias{pump}
\title{Pump-failure data}
\format{A matrix 10 rows and 2 columns.}
\source{
Table 3 of Gaver, D. P. and O'Muircheartaigh, I. G. (1987).
  See also Gelfand, A. E. and Smith, A. F. M. (1990).
}
\usage{
pump
}
\description{
Data on pump failures from Gaver, D. P. and O'Muircheartaigh, I. G. (1987).
The matrix \code{pump} has 10 rows and 2 columns.
Each row relates to a different pump system.
The first column contains the number of pump failures.
The second column contains the length of operating time, in
thousands of hours.
}
\references{
Gaver, D. P. and O'Muircheartaigh, I. G. (1987)
  Robust Empirical Bayes Analyses of Event Rates. \emph{Technometrics},
  \strong{29}, 1-15. \url{https://doi.org/10.1080/00401706.1987.10488178}

Gelfand, A. E. and Smith, A. F. M. (1990)
Sampling-Based Approaches to Calculating Marginal Densities.
\emph{Journal of the American Statistical Association}, \strong{85}(410),
398-409. \url{https://doi.org/10.1080/01621459.1990.10476213}
}
\keyword{datasets}
