% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_gamma.R
\name{sim_gamma_pois}
\alias{sim_gamma_pois}
\title{Simulate data from the gamma-Poisson model}
\usage{
sim_gamma_pois(J = 1, alpha = 1, beta = 1, exposure = 1)
}
\arguments{
\item{J}{An integer scalar. The number of groups.}

\item{alpha, beta}{Numeric vectors.  The parameters of the
gamma(\eqn{\alpha, \beta}) distribution for the Poisson rates
\eqn{\lambda1, ..., \lambdaJ}.}

\item{exposure}{A numeric scalar or a numeric vector of length \code{J}.
The exposures in groups 1, ..., \code{J}.}
}
\value{
A numeric matrix with 2 columns.  The first column, \code{y},
  contains the numbers of events, the second column, \code{exposure},
  the input \code{exposure}.
}
\description{
Simulates from the gamma-Poisson model described in \code{\link{hef}}.
}
\examples{
# Simulate data that are similar to the pump data
sim_data <- sim_gamma_pois(J = nrow(pump), exposure = pump[, "time"],
                           alpha = 1.2, beta = 2.2)
}
