% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_beta.R
\name{sim_pred_beta_binom}
\alias{sim_pred_beta_binom}
\title{Simulate from a beta-binomial posterior predictive distribution}
\usage{
sim_pred_beta_binom(theta_sim_vals, data, nrep)
}
\arguments{
\item{theta_sim_vals}{A numeric matrix with \code{nrow(data)} columns.
Each row of \code{theta_sim_vals} contains binomial success probabilities
simulated from their posterior distribution.}

\item{data}{A 2-column numeric matrix: the numbers of successes in column 1
and the corresponding numbers of trials in column 2.}

\item{nrep}{A numeric scalar.  The number of replications of the original
dataset simulated from the posterior predictive distribution.
If \code{nrep} is greater than \code{nrow(theta_sim_vals)} then
\code{nrep} is set equal to \code{nrow(theta_sim_vals)}.}
}
\value{
A numeric matrix with \code{nrep} rows.  Each row contains
  a draw from the posterior predictive distribution of the number of
  successes.
}
\description{
Simulates \code{nrep} draws from the posterior predictive distribution
of the beta-binomial model described in \code{\link{hef}}.
This function is called within \code{\link{hef}} when the argument
\code{nrep} is supplied.
}
\examples{
rat_res <- hef(model = "beta_binom", data = rat)
rat_sim_pred <- sim_pred_beta_binom(rat_res$theta_sim_vals, rat, 50)
}
