% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diagnostics.R
\name{summary.hef}
\alias{summary.hef}
\title{Summarizing hef objects}
\usage{
\method{summary}{hef}(object, ...)
}
\arguments{
\item{object}{an object of class "hef", a result of a call to \code{hef}.}

\item{...}{Additional arguments passed on to \code{\link[rust]{summary.ru}}.}
}
\description{
\code{summary} method for class "hef"
}
\examples{
# Binomial-beta model, rat data
rat_res <- hef(model = "binom_beta", data = rat)
summary(rat_res)
}
