% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bang-package.R
\docType{data}
\name{weight_gain}
\alias{weight_gain}
\title{Weight Gained by Rats}
\format{
A data frame with 40 rows and 3 columns.
  \itemize{
    \item Column 1, \code{source}: source of protein, a factor with levels
      \code{Beef} and \code{Cereal}.
    \item Column 2, \code{type}: amount of protein, a factor with levels
      \code{High} and \code{Low}.
    \item Column 3, \code{weightgain}: weight gained, in grams.
 }
}
\source{
D. J. Hand, A. D. Lunn, K. J. McConway, and E. Ostrowski (1994).
  \emph{A Handbook of Small Datasets}, Chapman and Hall/CRC, London.
}
\usage{
weight_gain
}
\description{
Data from an experiment to study weight gained by 10 rats fed on four different
diets, defined by a combination of the amount of protein (low and high)
and by the source of protein (beef and cereal).
}
\keyword{datasets}
