% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bang.R
\docType{package}
\name{bang}
\alias{bang}
\alias{bang-package}
\title{bang: Bayesian Analysis, No Gibbs}
\description{
Performs Bayesian analyses using some simple commonly-used models.
The multivariate generalized ratio-of-uniforms method is used to simulate
random samples from the required posterior distribution.
The user can either choose hyperparameter values of a default prior
distribution or specify their own prior distribution.
}
\details{
Currently three conjugate hierarchical models are available:
beta-binomial, gamma-Poisson and 1-way Analysis of Variance (ANOVA).
The function \code{\link{hef}} produces random posterior samples from for the
beta-binomial and gamma-Poisson models.  The function \code{\link{hanova1}}
does this for the 1-way Analysis of Variance (ANOVA).
The 'rust' package  <https://cran.r-project.org/package=rust> is used to
produce these samples.

See
vignette("bang-vignette", package = "bang") for a brief introduction
to the package and
vignette("revdbayes-anova-hef", package = "bang") and
vignette("revdbayes-anova-vignette", package = "bang") for illustrations
of the use of the \code{hef} and \code{hanova1} functions.
}
\references{
Northrop, P. J. (2017). rust: Ratio-of-Uniforms Simulation with
  Transformation. R package version 1.2.3.
  \url{https://cran.r-project.org/package=rust}.
}
\seealso{
The \code{\link[rust]{ru}} function in the \code{\link{rust}}
  package for details of the arguments that can be passed to \code{ru} via
  \code{hef} or \code{hanova1}.

\code{\link{set_user_prior}} to set a user-defined prior.
}
