% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hef.R
\name{hef}
\alias{hef}
\title{Hierarchical Exponential Family Model}
\usage{
hef(n = 1000, model = c("beta_binom", "gamma_pois"), data, ...,
  prior = "default", hpars = NULL, param = c("trans", "original"),
  init = NULL)
}
\arguments{
\item{n}{An integer scalar.  The size of the posterior sample required.}

\item{model}{A character string.  Abbreviated name for the
response-population distribution combination.
For a hierarchical normal model see \code{\link{hanova1}}
(hierarchical one-way analysis of variance (ANOVA)).}

\item{data}{A numeric matrix.  The format depends on \code{model}.
See \strong{Details}.}

\item{...}{Optional further arguments to be passed to
\code{\link[rust]{ru}}.}

\item{prior}{The log-prior for the parameters of the hyperprior
distribution.  If the user wishes to specify their own prior then
\code{prior} must be an object returned from a call to
\code{\link{set_user_prior}}.
Otherwise, \code{prior} is a character scalar giving the name of the
required in-built prior.
If \code{prior} is not supplied then a default prior is used.
See \strong{Details}.}

\item{hpars}{A numeric vector.  Used to set parameters (if any) in
an in-built prior.}

\item{param}{A character scalar.
If \code{param = "trans"} (the default) then the marginal posterior
of hyperparameter vector \eqn{\psi} is reparameterised in a way
designed to improve the efficiency of sampling from this posterior.
If \code{param = "original"} the original parameterisation is used.
The former tends to make the optimisations involved in the
ratio-of-uniforms algorithm more stable and to increase the probability
of acceptance, but at the expense of slower function evaluations.}

\item{init}{A numeric vector of length 2.  Optional initial estimates
for the search for the mode of the posterior density of the
hyperparameter vector \eqn{\psi}.}
}
\value{
An object (list) of class \code{"hef"}, which has the same
  structure as an object of class "ru" returned from \code{\link[rust]{ru}}.
  In particular, the columns of the \code{n}-row matrix \code{sim_vals}
  contain the simulated vaues of \eqn{\psi}.
  In addition this list contains the arguments \code{model}, \code{data},
  \code{prior} detailed above and an \code{n} by \eqn{J} matrix
  \code{theta_sim_vals}: column j contains the simulated values of
  (the first component of) \eqn{\thetaj}.
}
\description{
Samples from the posterior distribution of the parameters of
certain hierarchical exponential family models.  Conditional on
population-specific parameter vectors \eqn{\theta}1, ..., \eqn{\theta}J
the observed \emph{response} data \eqn{y}1, ..., \eqn{y}J within each
population are modelled as random samples from a distribution in an
exponential family. The population parameters \eqn{\theta}1, ...,
\eqn{\theta}J are modelled as random samples from a common
\emph{population distribution}, chosen to be conditionally conjugate
to the response distribution, with \emph{hyperparameter} vector \eqn{\phi}.
Conditionally on \eqn{\theta}1, ..., \eqn{\theta}J, \eqn{y}1, ..., \eqn{y}J
are independent of each other and are independent of \eqn{\psi}.
A \emph{hyperprior} is placed on \eqn{\psi}.  The user can either
choose parameter values of a default hyperprior or specify their own
hyperprior.
}
\details{
The \code{\link[rust]{ru}} function is used to draw a random sample
  from the marginal posterior of the hyperparameter vector \eqn{\psi}.
  Then, conditional on these values, population parameters are sampled
  directly from the conditional posterior density of
  \eqn{\theta}1, ..., \eqn{\theta}J given \eqn{\psi} and the data.

  We outline each \code{model}, specify the format of the
  \code{data}, give the default (log-)priors (up to an additive constant)
  and detail the choices of ratio-of-uniforms parameterisation
  \code{param}.

\strong{Beta-binomial:} For \eqn{j = 1, ..., J},
  \eqn{Yj | pj} are i.i.d binomial\eqn{(nj, pj)},
  where \eqn{pj} is the probability of success in group \eqn{j}
  and \eqn{nj} is the number of trials in group \eqn{j}.
  \eqn{pj} are i.i.d. beta\eqn{(\alpha, \beta)}, so
  and \eqn{\psi = (\alpha, \beta)}.
  \code{data} is a 2-column matrix: the numbers of successes in column 1
  and the corresponding numbers of trials in column 2.

\emph{Priors:}

\code{prior = "bda"} (the default):
\eqn{log \pi(\alpha, \beta) = - 2.5 log(\alpha + \beta),
  \alpha > 0, \beta > 0.} [See Section 5.3 of Gelman et al. (2014).]

\code{prior = "gamma"}: independent gamma priors on \eqn{\alpha}
and \eqn{\beta}, i.e.
\eqn{log \pi(\alpha, \beta) =
  (s1 - 1)log\alpha - r1 \alpha +
  (s2 - 1)log\beta - r2 \beta,  \alpha > 0, \beta > 0.}
where the respective shape (\eqn{s1}, \eqn{s2}) and rate
(\eqn{r1}, \eqn{r2}) parameters are specified using
\code{hpars} = \eqn{(s1, r1, s2, r2)}.  The default setting is
\code{hpars = c(1, 0.01, 1, 0.01).}

\emph{Parameterisations for sampling:}

 \code{param = "original"} is (\eqn{\alpha, \beta}),
 \code{param = "trans"} (the default) is
\eqn{\phi1 = logit(\alpha/(\alpha+\beta)) = log(\alpha/\beta),
  \phi2 = log(\alpha+\beta)}.
See Section 5.3 of Gelman et al. (2014).

\strong{Gamma-Poisson:} For \eqn{j = 1, ..., J},
  \eqn{Yj | \lambdaj} are i.i.d Poisson\eqn{(ej\lambdaj)},
  where
  \eqn{ej} is the \emph{exposure} in group \eqn{j}, based on the
  total length of observation time and/or size of the population at
  risk of the event of interest and \eqn{\lambdaj} is the mean number
  of events per unit of exposure.
  \eqn{\lambdaj} are i.i.d. gamma\eqn{(\alpha, \beta)}, so
  \eqn{\psi = (\alpha, \beta)}.
  \code{data} is a 2-column matrix: the counts \eqn{yj} in column 1
  and the corresponding exposures \eqn{ej} in column 2.

\emph{Priors:}

\code{prior = "gamma"} (the default): independent gamma priors
  on \eqn{\alpha} and \eqn{\beta}, i.e.
\eqn{log \pi(\alpha, \beta) =
  (s1 - 1)log\alpha - r1 \alpha +
  (s2 - 1)log\beta - r2 \beta,  \alpha > 0, \beta > 0.}
where the respective shape (\eqn{s1}, \eqn{s2}) and rate
(\eqn{r1}, \eqn{r2}) parameters are specified using
\code{hpars} = \eqn{(s1, r1, s2, r2)}.  The default setting is
\code{hpars = c(1, 0.01, 1, 0.01).}

\emph{Parameterisations for sampling:}

 \code{param = "original"} is (\eqn{\alpha, \beta}),
 \code{param = "trans"} (the default) is
\eqn{\phi1 = log(\alpha/\beta), \phi2 = log(\beta).}
}
\examples{
############################ Beta-binomial #################################

# ------------------------- Rat tumor data ------------------------------- #

# Default prior, sampling on (rotated) (log(mean), log(alpha + beta)) scale
rat_res <- hef(model = "beta_binom", data = rat)
plot(rat_res)
plot(rat_res, ru_scale = TRUE)
summary(rat_res)

# Default prior, sampling on (rotated) (alpha, beta) scale
rat_res <- hef(model = "beta_binom", data = rat, param = "original")
plot(rat_res)
plot(rat_res, ru_scale = TRUE)
summary(rat_res)

# To produce a plot akin to Figure 5.3 of Gelman et al. (2014) we
# (a) Use the same prior for (alpha, beta)
# (b) Don't use axis rotation (rotate = FALSE)
# (c) Plot on the scale used for ratio-of-uniforms sampling (ru_scale = TRUE)
# (d) Note that the mode is relocated to (0, 0) in the plot
rat_res <- hef(model = "beta_binom", data = rat, rotate = FALSE)
plot(rat_res, ru_scale = TRUE)
# This is the estimated location of the posterior mode
rat_res$f_mode

# User-defined prior, passing parameters
# (equivalent to prior = "gamma" with hpars = c(1, 0.01, 1, 0.01))
user_prior <- function(x, hpars) {
  return(dexp(x[1], hpars[1], log = TRUE) + dexp(x[2], hpars[2], log = TRUE))
}
user_prior_fn <- set_user_prior(user_prior, hpars = c(0.01, 0.01))
rat_res <- hef(model = "beta_binom", data = rat, prior = user_prior_fn)
plot(rat_res)
summary(rat_res)

############################ Gamma-Poisson #################################

# ------------------------ Pump failure data ------------------------------ #

pump_res <- hef(model = "gamma_pois", data = pump)
plot(pump_res)
plot(pump_res, ru_scale = TRUE)
summary(pump_res)
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S. Dunson, D. B.,
 Vehtari, A. and Rubin, D. B. (2014) \emph{Bayesian Data Analysis}.
 Chapman & Hall / CRC.
  \url{http://www.stat.columbia.edu/~gelman/book}
}
\seealso{
\code{\link{hanova1}} for hierarchical one-way analysis of
  variance (ANOVA).
}
