% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diagnostics.R
\name{plot.hef}
\alias{plot.hef}
\title{Plot diagnostics for a hef object}
\usage{
\method{plot}{hef}(x, y, type = c("hyper", "pop_sim", "pop_dens", "ru"),
  which_pop = 1, which_theta = 1, super = TRUE, ...)
}
\arguments{
\item{x}{an object of class "hef", a result of a call to \code{ru}.}

\item{y}{Not used.}

\item{type}{A character scalar.

  If \code{type = "hyper"} then the posterior sample of hyperparameter
  values is plotted using \code{\link[rust]{plot.ru}}.

  If \code{type = "ru"} then only the posterior sample generated using
  \code{\link[rust]{ru}} is plotted using \code{\link[rust]{plot.ru}}.

  If \code{type = "pop_sim"} then sample(s) of the population-specific
  parameter vector \eqn{\theta} are plotted.  The population(s) included
  are determined by \code{which_pop}.  If \code{which_pop} has length
  one a histogram is produced using \code{\link[graphics]{hist}}.
  Otherwise, pairwise scatter plots are produced using
  \code{\link[graphics]{pairs}}

  If \code{type = "pop_dens"} then posterior density estimates of
  population-specific parameter(s) are plotted. The population(s)
  included are determined by \code{which_pop}.
  The value of \code{super} determines whether or not these density
  estimates appear on one plot or in separate plots.}

\item{which_pop}{An integer vector.  If \code{type = "pop_sim"} or
\code{type = "pop_dens"} \code{which_pop} indicates which populations
to include in the plot.}

\item{which_theta}{An integer vector indicating which component of
of the population-specific parameter vector \eqn{\theta} to plot.
Only relevant if \eqn{\theta} has length greater than one.}

\item{super}{A logical scalar.  If \code{type = "pop_dens"} then
\code{super} indicates whether the density estimates should be
superimposed in one graph or appear in separate graphs.}

\item{...}{Additional arguments passed to \code{\link[rust]{plot.ru}},
\code{\link[graphics]{hist}} or \code{\link[graphics]{pairs}}.}
}
\description{
\code{plot} method for class "hef".  Calls \code{\link[rust]{plot.ru}}.
}
\examples{
# Binomial-beta model, rat data
rat_res <- hef(model = "binom_beta", data = rat)
# Hyperparameters alpha and beta
plot(rat_res)
plot(rat_res, ru_scale = TRUE)
# Population-specific parameter samples
plot(rat_res, type = "pop_sim")
plot(rat_res, type = "pop_sim", which_pop = c(1, 10))
}
\seealso{
\code{\link[rust]{plot.ru}} for arguments that may be passed
  via ...., in particular \code{ru_scale}.
}
