% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diagnostics.R
\name{plot.hef}
\alias{plot.hef}
\title{Plot diagnostics for a hef object}
\usage{
\method{plot}{hef}(x, y, ...)
}
\arguments{
\item{x}{an object of class "hef", a result of a call to \code{ru}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed on to \code{\link[rust]{plot.ru}}.}
}
\description{
\code{plot} method for class "hef".  Calls \code{\link[rust]{plot.ru}}.
}
\examples{
# Binomial-beta model, rat data
rat_res <- hef(model = "binom_beta", data = rat)
plot(rat_res)
plot(rat_res, ru_scale = TRUE)
}
\seealso{
\code{\link[rust]{plot.ru}} for the arguments that may be passed
  via ....
}
