% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_and_check_prior.R
\name{set_user_prior}
\alias{set_user_prior}
\title{Set a user-defined prior}
\usage{
set_user_prior(prior, model = c("binom_beta", "pois_gamma"), ...)
}
\arguments{
\item{prior}{An R function.}

\item{model}{A character string.  Abbreviated name for the
response-mixture combination.}

\item{...}{Further arguments giving the names and values of any
parameters involved in the function \code{prior}.}
}
\description{
Describe
}
