% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_beta.R
\name{sim_beta_binom}
\alias{sim_beta_binom}
\title{Simulate data from the beta-binomial model}
\usage{
sim_beta_binom(J = 1, size = 1, alpha = 1, beta = 1)
}
\arguments{
\item{J}{An integer scalar. The number of groups.}

\item{size}{A numeric scalar or a numeric vector of length \code{J}.
The number of trials in the groups 1, ..., \code{J}.  If \code{size} is a
scalar then \code{size} is used for all groups.}

\item{alpha, beta}{Numeric vectors.  The parameters of the
beta(\eqn{\alpha, \beta}) distribution for the binomial success
probabilities \eqn{p1, ..., pJ}.}
}
\value{
A numeric matrix with 2 columns.  The first column, \code{y},
  contains the numbers of successes, the second column, \code{n}, the
  numbers of trials.
}
\description{
Simulates from the beta-binomial model described in \code{\link{hef}}.
}
\examples{
# Simulate data that are similar to the rat data
sim_data <- sim_beta_binom(J = nrow(rat), size = rat[, "n"], alpha = 2.4,
                           beta = 14.3)
}
